requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
asgiref==3.9.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.5.0
debugpy==1.8.16
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.112.4
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.65.0
google-generativeai==0.8.6
googleapis-common-protos==1.70.0
gradio==4.43.0
gradio_client==1.3.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
instructor==1.14.5
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.13
langchain-google-genai==4.2.1
langchain-huggingface==1.2.0
langchain-openai==1.1.9
langchain-qdrant==1.1.0
langchain-tavily==0.2.17
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.3
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.1
openai==2.21.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
orjson==3.11.2
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.2
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
pip==25.2
platformdirs==4.3.8
pluggy==1.6.0
portalocker==3.2.0
posthog==6.5.0
prometheus_client==0.24.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.0.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.9.0
pydantic_core==2.23.2
pydantic-settings==2.13.0
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.11.0
PyMuPDF==1.27.2
pymupdf-layout==1.27.2
pymupdf4llm==1.27.2.1
pyodbc==5.2.0
pyparsing==3.2.3
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyprojroot==0.3.0
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
qdrant-client==1.15.1
ragas==0.4.3
rank-bm25==0.2.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
ruff==0.12.9
safetensors==0.6.2
scikit-learn==1.7.1
scikit-network==0.33.5
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.1.0
setuptools==57.4.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.38.6
sympy==1.14.0
tabulate==0.10.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
tomli==2.2.1
tomlkit==0.12.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
typer==0.16.0
typing_extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==16.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0